move each .ino into its own folder to be able to flash it...
vscode-arduino then asks which sketch to flash; also works in the Arduino IDE.
Tried platformio, but it does not allow to have several main files inside a project.
Would need to duplicate platformio.ini and have several folders in a workspace.
Also, the only dependency I have is not available in the official repository anyway.